# Improved performance of List.last/1