@josevalim 2020-09-19
Nice idea! Some concerns:

1. I believe the function now returns the wrong result (nil) for a list with exactly 8 elements (or a multiple of 8)
2. I wonder how JIT will affect these benchmarks - it may be that we are now optimizing for cases that will no longer be relevant in the future
3. If we do go ahead with this, you should add a comment explaining why we have such clause :)