@elbow-jason 2020-09-19
> I believe the function now returns the wrong result (nil) for a list with exactly 8 elements (or a multiple of 8)

OH THE BURNING SHAME! I don't know how I missed that. I'll put the last item of the 8 back in and see if the performance increase remains.

> I wonder how JIT will affect these benchmarks 

When JIT arrives, we should revisit this PR and benchmark with and without this change. No rush.

> it may be that we are now optimizing for cases that will no longer be relevant in the future.

That's true of everything that was ever built. :-) 

> add a comment explaining why we have such clause

I'll add it. :D