@elbow-jason 2020-09-19
> I'll put the last item of the 8 back in

I found a better way.