@michalmuskala 2020-09-19
I think we should compare that with the "naive" implementation combined with `@compile {:inline, last/1}` - AFAIK this should do some loop unrolling for recursive functions (which is what this PR does manually). I think it would be interesting to compare the results.