@elbow-jason 2020-09-19
@michalmuskala Like this? I'll run the benchmarks again.

```elixir
  @compile {:inline, [last: 1]}
  def last([]), do: nil
  def last([head]), do: head
  def last([_ | tail]), do: last(tail)
```