@michalmuskala 2020-09-19
Yes, exactly like that