@elbow-jason 2020-09-20
I made a repo so everyone could see the code, the results, and run their own benchmarks if desired.

https://github.com/elbow-jason/list_last_perf_experiment

In each run of benchmarks `list_last_inline` and `list_last8` were nearly always the same (or similar) in terms of performance and both were faster than `list_last_orig` (a copy of the current `List.last/1` implementation) except for the case where the input list is empty.

I think the best solution here would be to use the `list_last_inline` implementation instead of the `list_last8` implementation. The inline implementation is simpler, more maintainable, and requires fewer code changes.

Thoughts?