@josevalim 2020-09-20
:+1: for inline, as we do use it elsewhere too! Good call @michalmuskala!