@michalmuskala 2020-09-20
:+1: for the inline version from me as well. It's as maintainable as previous definition and relies on the compiler to make it fast instead of doing this by hand.