@sabiwara 2020-09-21
This was a very interesting PR!

Some naive question from somebody who doesn't know anything about inlining: the [erlang doc](http://erlang.org/doc/man/compile.html#inlining) says:
> Inlining does not necessarily improve running time. For example, inlining can increase Beam stack use, which probably is detrimental to performance for recursive functions.

The benchmarks seems to prove the opposite in this case, even if `List.last/1` is the perfect example of a recursive function... Is it because it is tail-recursive?

Side note: out of curiosity I also tried to benchmark it against `:lists.last/1` which is [has an interesting implementation](https://github.com/erlang/otp/blob/2d312773d995eb2177037c7530d12fb7c55f8308/lib/stdlib/src/lists.erl#L228), but it was typically slower for smaller lists and only marginally faster for big lists.