@michalmuskala 2020-09-21
I would be definitely opposed to blindly including the inlining directive, but given the measurements I think it makes sense. Inlining with recursion will be especially detrimental for functions that are not tail-recursive. Inlining a tail-recursive function should not increase the stack size if the inliner does its job well, since there can't be any stack around between recursive calls in the non-inlined implementation.

The erlang implementation is intersting - it checks for empty lists only on the first iteration `last/1`, but does not need to do this in the main loop of `last/2`. Our implementation checks for this on every iteration. I believe it would be beneficial to adopt a style similar to the erlang implementation, perhaps combined with the inline directive for even better gains.